requirements.txt
django
djangorestframework
dash
django_plotly_dash
dash-bootstrap-components
requests
pandas